with DTO and with inheritance

+++ src/main/java/com/alok/projectcars/controller/CarController.java
@@ -1,20 +1,19 @@
 package com.alok.projectcars.controller;
 
 import com.alok.projectcars.dao.model.Car;
-import com.alok.projectcars.domain.CarDomain;
+import com.alok.projectcars.dto.CarDto;
 import com.alok.projectcars.services.CarService;
 import com.alok.projectcars.services.DriverService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 import java.security.Principal;
 import java.util.List;
 
 @RestController
 @RequestMapping("/cars")
-public class CarController {
+public class CarController extends DriverController {
 
     @Autowired
     CarService carService;
@@ -22,18 +21,18 @@ public class CarController {
     @Autowired
     DriverService driverService;
 
-    @GetMapping("/all")
+    @GetMapping("/allCars")
     public ResponseEntity<?> getAllCars() {
 
-        List< ? extends Car> resultList;
+        List<? extends Car> resultList;
 
         resultList = carService.listAll();
 
         System.out.println(resultList.size());
-        if(resultList != null && resultList.size() > 0) {
+        if (resultList != null && resultList.size() > 0) {
             return new ResponseEntity<>(resultList, HttpStatus.OK);
-        } else  {
-           return new ResponseEntity<>("No content Found",HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>("No content Found", HttpStatus.NO_CONTENT);
         }
     }
 
@@ -48,23 +47,22 @@ public class CarController {
                                     @RequestParam(name = "manufacturer") String manufacturer, Principal principal) {
 
 
-            String roleOfCurrentUser = driverService.findByName(principal.getName()).getRole();
-            ResponseEntity<?> response;
-            if(roleOfCurrentUser.equals("ADMIN")) {
-                Car car = new Car(null,name,model,licensePlate,seatCount,convertible,rating,engineType,manufacturer,null,null,null);
+        String roleOfCurrentUser = driverService.findByName(principal.getName()).getRole();
+        ResponseEntity<?> response;
+        if (roleOfCurrentUser.equals("ADMIN")) {
+            Car car = new Car(null, name, model, licensePlate, seatCount, convertible, rating, engineType, manufacturer, null, null, null);
 
-                try {
-                    carService.save(car);
-                    response = new ResponseEntity<>(new CarDomain(car.getId(),car.getName()),HttpStatus.CREATED);
-                }
-                catch (Exception e) {
-                    response = new ResponseEntity<>("Error in saving the data",HttpStatus.INTERNAL_SERVER_ERROR);
-                }
-            } else  {
-                response = new ResponseEntity<>("It seems you have't access to this level ",HttpStatus.FORBIDDEN);
+            try {
+                carService.save(car);
+                response = new ResponseEntity<>(new CarDto(car.getId(), car.getName(), "Successfully Created!"), HttpStatus.CREATED);
+            } catch (Exception e) {
+                response = new ResponseEntity<>("Error in saving the data", HttpStatus.INTERNAL_SERVER_ERROR);
             }
+        } else {
+            response = new ResponseEntity<>("It seems you have't access to this level ", HttpStatus.FORBIDDEN);
+        }
 
-            return response;
+        return response;
     }
 
     @PutMapping
@@ -76,35 +74,49 @@ public class CarController {
                                         @RequestParam(name = "convertible") Boolean convertible,
                                         @RequestParam(name = "rating") Double rating,
                                         @RequestParam(name = "engineType") String engineType,
-                                        @RequestParam(name = "manufacturer") String manufacturer) {
-        Car car = carService.get(id);
+                                        @RequestParam(name = "manufacturer") String manufacturer,
+                                        Principal principal) {
 
-        car.setName(name);
-        car.setModel(model);
-        car.setLicensePlate(licensePlate);
-        car.setSeatCount(seatCount);
-        car.setConvertible(convertible);
-        car.setRating(rating);
-        car.setEngineType(engineType);
-        car.setManufacturer(manufacturer);
-        try {
-            carService.save(car);
-            return new ResponseEntity<>(new CarDomain(car.getId(),car.getName()),HttpStatus.OK);
+
+        String roleOfCurrentUser = driverService.findByName(principal.getName()).getRole();
+        if (roleOfCurrentUser.equals("ADMIN")) {
+            Car car = carService.get(id);
+            car.setName(name);
+            car.setModel(model);
+            car.setLicensePlate(licensePlate);
+            car.setSeatCount(seatCount);
+            car.setConvertible(convertible);
+            car.setRating(rating);
+            car.setEngineType(engineType);
+            car.setManufacturer(manufacturer);
+            try {
+                carService.save(car);
+                return new ResponseEntity<>(new CarDto(car.getId(), car.getName(), "Successfully record updated!"), HttpStatus.OK);
+            } catch (Exception e) {
+                return new ResponseEntity<>("Error in saving the data", HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+        } else {
+            return new ResponseEntity<>("It seems you have't access to this level ", HttpStatus.FORBIDDEN);
         }
-        catch (Exception e) {
-           return new ResponseEntity<>("Error in saving the data",HttpStatus.INTERNAL_SERVER_ERROR);
-        }
+
     }
 
     @DeleteMapping
-    public ResponseEntity<?> deleteCar(@RequestParam(name = "id") Long id) {
+    public ResponseEntity<?> deleteCar(@RequestParam(name = "id") Long id,
+                                        Principal principal) {
         ResponseEntity<?> response;
-        try {
-            carService.delete(id);
-            response = new ResponseEntity<>("Successfully deleted car with ID : "+id,HttpStatus.NO_CONTENT);
-        } catch (Exception e) {
-            response = new ResponseEntity<>("Error in deleting the data",HttpStatus.INTERNAL_SERVER_ERROR);
+        String roleOfCurrentUser = driverService.findByName(principal.getName()).getRole();
+        if (roleOfCurrentUser.equals("ADMIN")) {
+            try {
+                carService.delete(id);
+                response = new ResponseEntity<>("Successfully deleted car with ID : " + id, HttpStatus.NO_CONTENT);
+            } catch (Exception e) {
+                response = new ResponseEntity<>("Error in deleting the data", HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+        } else  {
+            return new ResponseEntity<>("It seems you have't access to this level ", HttpStatus.FORBIDDEN);
         }
+
         return response;
     }
 }
